requirements.txt
Django==2.2
django-elasticsearch-dsl==0.5.1
django-elasticsearch-dsl-drf==0.17.6
django-nine==0.1.13
djangorestframework==3.9.2
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
ipaddress==1.0.22
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
